Move valid_node/valid_non_root into HugrView (no uses yet)
