+++ src/hugr/views.rs
@@ -18,7 +18,7 @@ use itertools::{Itertools, MapInto};
 use portgraph::dot::{DotFormat, EdgeStyle, NodeStyle, PortStyle};
 use portgraph::{multiportgraph, LinkView, MultiPortGraph, PortView};
 
-use super::{Hugr, NodeMetadata, NodeType};
+use super::{Hugr, HugrError, NodeMetadata, NodeType};
 use crate::ops::handle::NodeHandle;
 use crate::ops::{FuncDecl, FuncDefn, OpName, OpTag, OpType, DFG};
 use crate::types::{EdgeKind, FunctionType};
@@ -79,6 +79,30 @@ pub trait HugrView: sealed::HugrInternals {
 
     /// Returns whether the node exists.
     fn contains_node(&self, node: Node) -> bool;
+
+    /// Validates that a node is valid in the graph.
+    ///
+    /// Returns a [`HugrError::InvalidNode`] otherwise.
+    #[inline]
+    fn valid_node(&self, node: Node) -> Result<(), HugrError> {
+        match self.contains_node(node) {
+            true => Ok(()),
+            false => Err(HugrError::InvalidNode(node)),
+        }
+    }
+
+    /// Validates that a node is a valid root descendant in the graph.
+    ///
+    /// To include the root node use [`HugrView::valid_node`] instead.
+    ///
+    /// Returns a [`HugrError::InvalidNode`] otherwise.
+    #[inline]
+    fn valid_non_root(&self, node: Node) -> Result<(), HugrError> {
+        match self.root() == node {
+            true => Err(HugrError::InvalidNode(node)),
+            false => self.valid_node(node),
+        }
+    }
 
     /// Returns the parent of a node.
     fn get_parent(&self, node: Node) -> Option<Node>;
